requirements.txt
pillow
pyyaml
praw
pytesseract
